Update SampleAngularPage.page
This page provides an example on how to leverage AngularJS and Bootstrap using remote objects and JS remoting within a Visualforce page without a view state.  The page displays a table of contacts that is searchable, sortable, and paginated.  The pagination is handled in a VF component in conjunction with an Angular factory.